test: add delete training plan test

+++ src/module/training-plan/core/service/training-plan-management.service.ts
@@ -33,6 +33,9 @@ export class TrainingPlanManagementService {
   }
 
   async delete(id: string) {
+    const exists = await this.exists(id);
+    if (!exists) throw new NotFoundException(`training plan with id ${id} not found`);
+
     return await this.trainingPlanRepository.deleteTrainingPlan(id);
   }
